Large file update
In short, lines 12-22 had to be moves some and I added for_loop and def_scan port function to allow the scan port function to communicate with the tkinter gui. Overall, took a good bit of research, the AI portion in my Visusal studio and we got it. We need to add a "cancel scan" input function... it got hung a few times and I had to ctrl alt delete